feat(ImproveLoading): use custom component and stack the login routes
