+++ package.json
@@ -105,6 +105,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.45.4",
+    "react-icons": "^5.5.0",
     "react-slick": "^0.30.3",
     "sharp": "^0.32.6",
     "slick-carousel": "^1.8.1",
