+++ src/app/(payload)/admin/login/custom.scss
@@ -1,0 +1,229 @@
+@import '~@payloadcms/ui/scss';
+
+.wrapper-outsides {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  min-height: 100vh;
+  background-color: white;
+
+  .form-wrapper {
+    max-width: 600px;
+    width: 100%;
+    
+    .logo-section {
+      display: flex;
+      align-items: center;
+      gap: 12px;
+      margin-bottom: 24px;
+      justify-content: center;
+      text-align: center;
+
+      img {
+        height: 20px;
+        width: 20px;
+      }
+
+      p {
+        font-size: 18px;
+        font-weight: 600;
+        color: #111827;
+      }
+    }
+
+    .welcome-text {
+      text-align: center;
+      margin-bottom: 24px;
+      color: #374151;
+    }
+
+    .login-form-container {
+      padding: 48px 32px;
+      background: white;
+      border-radius: 8px;
+      box-shadow: 0 1px 3px rgba(0,0,0,0.1);
+      border: 1px solid #e5e7eb;
+
+      .login__form__inputWrap {
+        margin-bottom: 24px;
+
+        .field-type {
+          margin-bottom: 16px;
+
+          label {
+            display: block;
+            font-size: 14px;
+            font-weight: 500;
+            margin-bottom: 8px;
+            color: #374151;
+
+            .required {
+              color: #ef4444;
+            }
+          }
+
+          input {
+            width: 100%;
+            height: 48px;
+            padding: 0 16px;
+            border: 1px solid #d1d5db;
+            border-radius: 6px;
+            background: white;
+            color: #111827;
+            &:-webkit-autofill,
+            &:-webkit-autofill:hover,
+            &:-webkit-autofill:focus {
+              -webkit-box-shadow: 0 0 0 30px white inset !important;
+              -webkit-text-fill-color: #111827 !important;
+              transition: background-color 5000s ease-in-out 0s;
+            }
+
+            &:focus {
+              outline: none;
+              border-color: #2563eb;
+              box-shadow: 0 0 0 2px rgba(37,99,235,0.1);
+            }
+          }
+        }
+
+        .forgot-password {
+          color: #6b7280;
+          text-decoration: none;
+          font-size: 14px;
+          &:hover {
+            text-decoration: underline;
+          }
+        }
+      }
+
+      .form-submit {
+        margin-bottom: 16px;
+
+        button {
+          width: 100%;
+          height: 48px;
+          background: #2563eb;
+          color: white;
+          border: none;
+          border-radius: 6px;
+          font-weight: 500;
+          cursor: pointer;
+
+          &:hover {
+            background: #1d4ed8;
+          }
+
+          &:disabled {
+            opacity: 0.5;
+            cursor: not-allowed;
+          }
+        }
+      }
+      /* hide for a while */
+      // .google-button {
+      //   width: 100%;
+      //   height: 48px;
+      //   background: white;
+      //   border: 1px solid #d1d5db;
+      //   border-radius: 6px;
+      //   display: flex;
+      //   align-items: center;
+      //   justify-content: center;
+      //   gap: 8px;
+      //   color: #374151;
+      //   font-weight: 500;
+      //   cursor: pointer;
+
+      //   &:hover {
+      //     background: #f9fafb;
+      //   }
+      // }
+    }
+  }
+}
+.toast-message {
+  position: fixed;
+  bottom: 24px;
+  right: 24px;
+  padding: 16px 24px;
+  background: var(--theme-error-500);
+  color: white;
+  border-radius: 4px;
+  animation: slideIn 0.3s ease-out;
+  z-index: 1000;
+}
+
+@keyframes slideIn {
+  from {
+    transform: translateX(100%);
+    opacity: 0;
+  }
+  to {
+    transform: translateX(0);
+    opacity: 1;
+  }
+}
+
+.form-fade-in {
+  animation: fadeIn 0.5s ease-out;
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+    transform: translateY(20px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+.loading-overlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: rgba(26, 26, 26, 0.7);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  backdrop-filter: blur(2px);
+  animation: fadeIn 0.2s ease-out;
+}
+
+.loading-spinner {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1rem;
+
+  svg {
+    animation: spin 1s linear infinite;
+  }
+
+  span {
+    font-size: 14px;
+    color: #fff;
+  }
+}
+
+
+.btn {
+  &:disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+    pointer-events: none;
+    background: var(--theme-elevation-400) !important;
+  }
+}
+
+@keyframes spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
