--- src/app/(payload)/custom.scss
@@ -1,5 +1,0 @@
-/* Hide the login form default */
-.login__form,
-.login__form form {
-  display: none !important;
-}
